Add ws allBookTickersEvent and tickerEvent
