+++ src/main/java/com/binance/api/client/BinanceApiWebSocketClient.java
@@ -14,8 +14,8 @@ public interface BinanceApiWebSocketClient extends Closeable {
     /**
      * Open a new web socket to receive {@link DepthEvent depthEvents} on a callback.
      *
-     * @param symbols   market (one or coma-separated) symbol(s) to subscribe to
-     * @param callback  the callback to call on new events
+     * @param symbols  market (one or coma-separated) symbol(s) to subscribe to
+     * @param callback the callback to call on new events
      * @return a {@link Closeable} that allows the underlying web socket to be closed.
      */
     Closeable onDepthEvent(String symbols, BinanceApiCallback<DepthEvent> callback);
@@ -23,9 +23,9 @@ public interface BinanceApiWebSocketClient extends Closeable {
     /**
      * Open a new web socket to receive {@link CandlestickEvent candlestickEvents} on a callback.
      *
-     * @param symbols   market (one or coma-separated) symbol(s) to subscribe to
-     * @param interval  the interval of the candles tick events required
-     * @param callback  the callback to call on new events
+     * @param symbols  market (one or coma-separated) symbol(s) to subscribe to
+     * @param interval the interval of the candles tick events required
+     * @param callback the callback to call on new events
      * @return a {@link Closeable} that allows the underlying web socket to be closed.
      */
     Closeable onCandlestickEvent(String symbols, CandlestickInterval interval, BinanceApiCallback<CandlestickEvent> callback);
@@ -33,8 +33,8 @@ public interface BinanceApiWebSocketClient extends Closeable {
     /**
      * Open a new web socket to receive {@link AggTradeEvent aggTradeEvents} on a callback.
      *
-     * @param symbols   market (one or coma-separated) symbol(s) to subscribe to
-     * @param callback  the callback to call on new events
+     * @param symbols  market (one or coma-separated) symbol(s) to subscribe to
+     * @param callback the callback to call on new events
      * @return a {@link Closeable} that allows the underlying web socket to be closed.
      */
     Closeable onAggTradeEvent(String symbols, BinanceApiCallback<AggTradeEvent> callback);
@@ -49,21 +49,38 @@ public interface BinanceApiWebSocketClient extends Closeable {
     Closeable onUserDataUpdateEvent(String listenKey, BinanceApiCallback<UserDataUpdateEvent> callback);
 
     /**
-     * Open a new web socket to receive {@link AllMarketTickersEvent allMarketTickersEvents} on a callback.
+     * Open a new web socket to receive {@link TickerEvent tickerEvents} on a callback.
+     *
+     * @param symbols  market (one or coma-separated) symbol(s) to subscribe to
+     * @param callback the callback to call on new events
+     * @return a {@link Closeable} that allows the underlying web socket to be closed.
+     */
+    Closeable onTickerEvent(String symbols, BinanceApiCallback<TickerEvent> callback);
+
+    /**
+     * Open a new web socket to receive {@link List<TickerEvent> allMarketTickersEvents} on a callback.
      *
      * @param callback the callback to call on new events
      * @return a {@link Closeable} that allows the underlying web socket to be closed.
      */
-    Closeable onAllMarketTickersEvent(BinanceApiCallback<List<AllMarketTickersEvent>> callback);
+    Closeable onAllMarketTickersEvent(BinanceApiCallback<List<TickerEvent>> callback);
 
     /**
      * Open a new web socket to receive {@link BookTickerEvent bookTickerEvents} on a callback.
      *
-     * @param symbols   market (one or coma-separated) symbol(s) to subscribe to
-     * @param callback  the callback to call on new events
+     * @param symbols  market (one or coma-separated) symbol(s) to subscribe to
+     * @param callback the callback to call on new events
      * @return a {@link Closeable} that allows the underlying web socket to be closed.
      */
     Closeable onBookTickerEvent(String symbols, BinanceApiCallback<BookTickerEvent> callback);
+
+    /**
+     * Open a new web socket to receive {@link TickerEvent allBookTickersEvents} on a callback.
+     *
+     * @param callback the callback to call on new events
+     * @return a {@link Closeable} that allows the underlying web socket to be closed.
+     */
+    Closeable onAllBookTickersEvent(BinanceApiCallback<BookTickerEvent> callback);
 
     /**
      * @deprecated This method is no longer functional. Please use the returned {@link Closeable} from any of the other methods to close the web socket.
